Name service only with name_service_true + multi-threading
1. The name service is now only activated when the URL contains
&name_service=true . Before it was either one or off for all queries,
which did not make too much sense. Note that it's usually only the
"final" queries sent to the backend for which we may want the name
service. And not for "internal" queries like those issued for the SPARQL
autocompletion or by the Maps UI.

2. Made the server multi-threaded. So far it was single-threaded, which
means it stalled, for example, when a hard autocompletion query was
issued and then the query modified to launch a simpler query. The code
now requires Python >= 3.7 though, but that should be standard pretty
soonC. On Ubuntu 18.04, the standard version is still 3.6.x though. But
it's easy to install Python 3.7.

3. Made several improvements to the log. TODO: add a prefix for each
line in the log, so that it's clear to which request it belongs. Right
now, the log is rather unreadable when ineeded multiple queries are
processed concurrently.