Adapt to new cmd=... and error codes
1. The various options for the URL parameter cmd now have a dash. For
example, it's now clear-cache instead of clearcache.

2. QLever now returns a status code of 400 (Bad Request) for queries
that could not be processed and will return other error codes in the
future. So far, for status codes other than 200, the proxy returned
a generic "proxy error" to the client (i.e., the QLever UI). With this
fix, the responses are now passed through and the status code is maintained
(before it was always set to 200, independent of the status code
returned by the backend).